fix #1 [2/2] advance():
The current approach I took was to store those so called Imlib_Image
pointer (aka "IMAGES") and store them in the struct Image. This is not
smart when you got full hd images.. This also wastes resources for
unused images. This should only be done with thumbnails, croping them in
a small and easy to handle size (like the freedesktop standard uses).

Solution:
	free the image that exist in the imlib context, in other words:
	free the image that we alrd loaded in loadimgs() and are no
	longer in use.